feat: add user controller and tests + fix delete functions in BookingService and SeatService

+++ backend/src/main/java/com/rayan/cinemaapi/service/BookingService.java
@@ -55,6 +55,10 @@ public class BookingService {
     }
 
     public void deleteBooking(Long id) {
+        if (!bookingRepository.existsById(id)) {
+            throw new EntityNotFoundException("Booking", id);
+        }
+
         bookingRepository.deleteById(id);
     }
 }
